Generate digest nonce random data with C++11 generator
This should also resolve the bug 3526 properly since the new generator
is guaranteed to produce new random values every time with a higher
quality of randomness than rand() which was being used indirectly
before.

+++ src/auth/digest/Config.h
@@ -32,7 +32,7 @@ struct _digest_nonce_data {
     time_t creationtime;
     /* in memory address of the nonce struct (similar purpose to an ETag) */
     digest_nonce_h *self;
-    long randomdata;
+    uint32_t randomdata;
 };
 
 /* the nonce structure we'll pass around */
